improoved go_manager update (less list copying). started with level dependend bugs.

+++ go_manager.py
@@ -33,29 +33,26 @@ class GoManager(object):
         for go in self._go_list:
             go.update(delta)
 
-        #get all lists of different gameobjects
-        snips = [go for go in self._go_list if isinstance(go, snippets.Snippet)]
-        excepties = [go for go in self._go_list if isinstance(go, exceptions.Excepties)]
         bugs_ = [go for go in self._go_list if isinstance(go, bugs.Bugs)]
 
-        for ex in excepties:
-            player_.try_catch(ex)
+        for go in self._go_list:
+            if isinstance(go, exceptions.Excepties):
+                player_.try_catch(go)
 
-        for s in snips:
-            s_surface = s.surface
-            s_x, s_y = s.pos
-            for b in bugs_:
-                if collision.collide(s_surface, s_x, s_y, b.surface, *b.pos):
-                    s.delete()
-                    b.delete()
-                    status_.inc_point()
-                    self._sound.play(sound.Sounds.KILL)
+            elif isinstance(go, snippets.Snippet):
+                s = go
+                s_surface = s.surface
+                s_x, s_y = s.pos
+                for b in bugs_:
+                    if not b._delete_me:
+                        if collision.collide(s_surface, s_x, s_y, b.surface, *b.pos):
+                            self._sound.play(sound.Sounds.KILL)
+                            s.delete()
+                            b.delete()
+                            status_.inc_point()
+                            break
 
         # takes only those go to the list which are go._delete_me == true
         self._go_list = [go for go in self._go_list if not go._delete_me]
 
-
-
-if __name__ == '__main__':
-    print([i for i in range(0, 10) ])
